Fix a tile 404 stopping the seeding from completing
When seeding tiles, if one cannot be found (i.e. it 404’s) then the script was never continuing and it was stuck seeding forever.

This is because it only continues onto the next tile once the tile has downloaded, or it has been found in the cache. I’ve added in some logic to skip the tile (i.e. don’t cache the tile) if it errors when loading - so seeding can continue.

Tiles 404 a lot when using retina tiles (detectRetina: true). Fundamentally, theres a bug with caching retina tiles. But in lieu of fixing that, we’ve made sure that it never gets stuck.